add test platform

+++ inc/lib/string.h
@@ -1,16 +1,19 @@
 #pragma once
+#include <stdint.h>
+#include <stddef.h>
+#include <snprintf/snprintf.h>
 
 #ifdef __cplusplus
 class String {
    public:
-    static long Length(const char* str);
+    static size_t Length(const char* str);
     static void Copy(const char* src, char* dest);
 };
 
 extern "C" {
 #endif
 
-static long strlen(const char* str);
+static size_t strlen(const char* str);
 static char* strcpy(char* dest, const char* src);
 
 #ifdef __cplusplus
